Angular 9 SSR integration

+++ src/app/shared/services/storage.service.ts
@@ -1,36 +1,44 @@
-import { Injectable } from '@angular/core';
+import { PLATFORM_ID, Inject, Injectable } from '@angular/core';
+import { isPlatformServer } from '@angular/common';
 
 import { AuthUser } from '../../models/auth-user.model';
 
 @Injectable()
 export class StorageService {
-  constructor() {}
+  constructor(@Inject(PLATFORM_ID) private platformId: object) {}
 
   getUser(): AuthUser {
+    if (isPlatformServer(this.platformId)) return null;
     return JSON.parse(window.localStorage.getItem('authUser'));
   }
 
   saveUser(user: AuthUser) {
+    if (isPlatformServer(this.platformId)) return;
     window.localStorage.setItem('authUser', JSON.stringify(user));
   }
 
   getAccessToken(): string {
+    if (isPlatformServer(this.platformId)) return null;
     return window.localStorage['accessToken'];
   }
 
   saveAccessToken(accessToken: string) {
+    if (isPlatformServer(this.platformId)) return;
     window.localStorage['accessToken'] = accessToken;
   }
 
   getRefreshToken(): string {
+    if (isPlatformServer(this.platformId)) return null;
     return window.localStorage['refreshToken'];
   }
 
   saveRefreshToken(refreshToken: string) {
+    if (isPlatformServer(this.platformId)) return;
     window.localStorage['refreshToken'] = refreshToken;
   }
 
   getACL(): any {
+    if (isPlatformServer(this.platformId)) return null;
     if (window.localStorage.getItem('acl')) {
       return JSON.parse(window.localStorage.getItem('acl'));
     }
@@ -38,6 +46,7 @@ export class StorageService {
   }
 
   saveACL(acl: any) {
+    if (isPlatformServer(this.platformId)) return;
     window.localStorage.setItem('acl', JSON.stringify(acl));
   }
 
@@ -50,11 +59,13 @@ export class StorageService {
   }
 
   getLanguageCode(): string {
+    if (isPlatformServer(this.platformId)) return null;
     const index = ['en', 'bd'].indexOf(window.localStorage['languageCode']);
     return index >= 0 ? window.localStorage['languageCode'] : 'en';
   }
 
   setLanguageCode(lanCode: string) {
+    if (isPlatformServer(this.platformId)) return;
     const index = ['en', 'bd'].indexOf(lanCode);
 
     if (index >= 0) {
@@ -65,6 +76,7 @@ export class StorageService {
   }
 
   destroyAll() {
+    if (isPlatformServer(this.platformId)) return;
     const language = this.getLanguageCode();
     window.localStorage.clear();
     this.setLanguageCode(language);
